# apps/mobile/contracts/openapi.mobile.yaml
openapi: 3.0.3
info:
  title: PawfectMatch Mobile API
  description: API endpoints used by the PawfectMatch mobile application
  version: 1.0.0
  contact:
    name: Mobile Engineering Team
    email: mobile@pawfectmatch.com
servers:
  - url: https://api.pawfectmatch.com
    description: Production server
  - url: https://api-staging.pawfectmatch.com
    description: Staging server
  - url: http://localhost:3001
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /uploads:
    post:
      summary: Register upload and start processing
      tags: [Uploads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - petId
                - s3Key
              properties:
                petId:
                  type: string
                s3Key:
                  type: string
      responses:
        '200':
          description: Upload registered for processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Upload'

    get:
      summary: Get uploads for user
      tags: [Uploads]
      parameters:
        - name: petId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, ai_scanning, approved, rejected]
      responses:
        '200':
          description: Uploads retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Upload'
  /uploads/{uploadId}:
    get:
      summary: Get upload status
      tags: [Uploads]
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Upload status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Upload'

  # Verification
...
  /chats/{chatId}/messages:
    get:
      summary: Get chat messages
      tags: [Chat]
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
        - name: before
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Messages retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Send message
      tags: [Chat]
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 1000
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Message'

  # Map
...
components:
  schemas:
...
    Upload:
      type: object
      properties:
        id:
          type: string
        petId:
          type: string
        status:
          type: string
          enum: [pending, ai_scanning, approved, rejected]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        s3Key:
          type: string
        moderationResult:
          type: object
          nullable: true
          properties:
            status:
              type: string
              enum: [approved, rejected]
            reason:
              type: string
              enum: [safe, unsafe_content, duplicate]
            confidence:
              type: number
              minimum: 0
              maximum: 1
            categories:
              type: array
              items:
                type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Message:
      type: object
      properties:
        id:
          type: string
        chatId:
          type: string
        senderId:
          type: string
        text:
          type: string
        timestamp:
          type: string
          format: date-time
        read:
          type: boolean
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT